Moved helper functions to helpers.go
